backend/pyproject.toml
[project]
name = "backend"
version = "0.0.0"
description = "FastAPI Backend boilerplate"
authors = [{ name = "mrdsx" }]
readme = "README.md"
license = "LICENSE"
requires-python = "3.14"
dependencies = [
  "fastapi (>=0.122.0,<0.123.0)",
  "uvicorn (>=0.38.0,<0.39.0)",
  "python-dotenv (>=1.2.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
